Session improvements

+++ src/store/session.js
@@ -4,23 +4,24 @@ import firebase from 'firebase/app'
 import 'firebase/auth'
 import * as database from 'database'
 import player from 'store/player'
-import requests from 'store/requests'
 import errors from 'store/errors'
+import screen from 'store/screen'
 import storage from 'store/storage'
+import requests from 'store/requests'
 import appIsReady from 'store/appIsReady'
 
 export default {
   addAuthenticationListener() {
     firebase.auth().onAuthStateChanged(async authData => {
       requests.start('authStateChange')
-      console.log({ authData })
 
       // If user is signing in...
       if (authData && get(player) === null) {
         const domain = authData.email.split('@').pop().split('.').slice(-2)[0]
+        const loggedInManually = authData.providerData[0].providerId === 'password'
 
         // Don't let non-Monedo users in
-        if (!['monedo', 'kreditech'].includes(domain)) {
+        if (!loggedInManually && !['monedo', 'kreditech'].includes(domain)) {
           database.signOut()
           firebase.auth().currentUser.delete()
           requests.stop('authStateChange')
@@ -36,7 +37,6 @@ export default {
         // If the data is not found, create it using data from the authentication object
         if (!playerData) {
           const emailSlug = authData.email.split('@')[0].split('+')[0].replace(/[._-]/g, '')
-
           const characterTemplates = await database.get('characterTemplates')
 
           const template = {
@@ -46,10 +46,11 @@ export default {
 
           playerData = {
             id: authData.uid,
-            name: authData.displayName,
+            name: authData.displayName || emailSlug,
             email: authData.email,
             emailSlug,
-            photoUrl: authData.photoURL,
+            // TODO:L remove kitten
+            photoUrl: authData.photoURL || 'http://placekitten.com/200/300',
             ...template
           }
         }
@@ -67,7 +68,6 @@ export default {
 
         storage.save('signedIn', true)
         player.set(playerData)
-        console.log({ playerData })
         requests.stop('signIn')
       }
 
@@ -85,37 +85,19 @@ export default {
   signIn() {
     if (get(player) || get(requests).signIn) return
 
-    console.log('SIGN IN')
     errors.hide('signIn')
+    errors.hide('wrongEmailDomain')
     requests.start('signIn')
-
-    // NOTE: real code
     database.signIn().catch(error => {
       errors.show('signIn', error)
       requests.stop('signIn')
     })
-
-    // KILLME: mocks
-    // requests.start('authStateChange')
-    // storage.save('signedIn', true)
-    // player.set({
-    //   id: 'M0ck3d1d',
-    //   name: 'Robert Kirsz',
-    //   email: 'robert.kirsz@gmail.com',
-    //   emailSlug: 'robertkirsz',
-    //   photoUrl: 'https://lh3.googleusercontent.com/a-/AAuE7mAPnMxV4UEnKcDQ8I5JN8I3ScdvZEqEbZnp8Hta-Ig'
-    // })
-    // requests.stop('signIn')
-    // appIsReady.set(true)
-    // requests.stop('authStateChange')
   },
   manualSignIn(email, password) {
     if (get(player) || get(requests).signIn) return
 
-    console.log('MANUAL SIGN IN')
     errors.hide('signIn')
     requests.start('signIn')
-
     database.manualSignIn(email, password).catch(error => {
       errors.show('signIn', error)
       requests.stop('signIn')
@@ -126,22 +108,14 @@ export default {
 
     if (!playerData || get(requests).signOut) return
 
-    console.log('SIGN OUT')
     screen.close('MENU')
     errors.hide('signOut')
     requests.start('signOut')
 
-    // NOTE: real code
     await database
-      .update(`players/${playerData.id}`, {
-        ...playerData,
-        isOnline: false
-      })
-      .catch(error => {
-        errors.show('updateUserInSignOut', error)
-      })
+      .update(`players/${playerData.id}`, { ...playerData, isOnline: false })
+      .catch(error => errors.show('updateUserInSignOut', error))
 
-    // NOTE: real code
     database
       .signOut()
       // TODO: Don't know if needed, we do saveToLocalStorage('signedIn', false) in addAuthenticationListener signOut case
@@ -150,14 +124,5 @@ export default {
         errors.show('signOut', error)
         requests.stop('signOut')
       })
-
-    // KILLME: mocks
-    // requests.start('authStateChange')
-    // storage.save('signedIn', false)
-    // player.set(null)
-    // requests.stop('signOut')
-    // appIsReady.set(true)
-    // requests.stop('authStateChange')
-    // storage.clear()
   }
 }
